Upload a file by specifying its raw text content and metadata values without having to upload a file.

Upload a file by specifying its raw text content and metadata values without having to upload a file.

Usage

documents_raw_text(title, text, ...)

Arguments

title

file title

text

raw text of file

...

other metadata with key and values

Value

The document metadata